backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-google-genai
google-generativeai
langchain-community
sentence-transformers
langchain-pinecone
langchain-text-splitters
pymongo
pinecone
pydantic-settings
python-dotenv